requirements/dev.txt
# Local development dependencies go here
-r base.txt

# django dev tools
django-debug-toolbar
# Until https://github.com/jazzband/django-silk/commit/5bf4c4975b60d26fa873c8dbd265ddc1144e3328 goes live need to use old version
django-silk==3.0.1

# python linting
flake8
flake8-colors
flake8-isort
flake8-print

# test dependencies
#coverage
#mock

# python module dependencies
pip-autoremove

# python import sorter
isort

# show pip dependency tree
pipdeptree

# tox (test with multiple python/django versions)
#tox

# needed for django-extensions' runserver_plus
werkzeug

